Add ability to set secrets, mode, and ownership
This adds the ability to specify a `secrets` attribute set, where each key is the name of a variable (specified by `${var}`), and the value is the path to a file which contains the value to use to replace that variable.

It also allows specifying the mode, owner, and group for the resulting file.

Secrets are never revealed in the Nix store.